+++ CoC.Frontend/Encounters/Forest/CorruptedGrowthEncounter.cs
@@ -88,7 +88,7 @@ namespace CoC.Frontend.Encounters.Forest
 			throw new InDevelopmentExceptionThatBreaksOnRelease();
 		}
 
-		private void runCombatEncounter(float baseHealth = 1.0f)
+		private void runCombatEncounter(double baseHealth = 1.0f)
 		{
 			throw new InDevelopmentExceptionThatBreaksOnRelease();
 		}
